LMS-saas-django-main/requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.0.0
django-ckeditor-5==0.2.13
django-debug-toolbar==6.0.0
django-filter==25.1
django-tenants==3.8.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
Faker==37.5.3
idna==3.10
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-decouple==3.8
redis==6.4.0
requests==2.32.4
sqlparse==0.5.3
tzdata==2025.2
urllib3>=1.21.1,<1.27
gunicorn==23.0.0
boto3==1.38.20
django-storages==1.14.6
django-tinymce==4.1.0
django-cors-headers==4.7.0
django-silk==5.4.2
drf-writable-nested==0.7.2
openpyxl==3.1.5
sendgrid==6.12.5
google-generativeai==0.3.2
